+++ src/skimage/restoration/unwrap_2d_ljmu.h
@@ -4,7 +4,7 @@ void unwrap2D(
         double *wrapped_image,
         double *UnwrappedImage,
         unsigned char *input_mask,
-        int n_j, int n_i,
+        intptr_t n_j, intptr_t n_i,
         int wrap_around_j, int wrap_around_i,
         bitgen_t* bitgen_state
         );
